+++ include/global.fieldmap.h
@@ -218,7 +218,7 @@ struct ObjectEvent
     u8 directionOverwrite:4;
     u8 directionSequenceIndex;
     u8 playerCopyableMovement; // COPY_MOVE_*
-    u8 spriteId;
+    u32 spriteId;
 };
 
 struct ObjectEventGraphicsInfo
@@ -319,8 +319,8 @@ struct PlayerAvatar
     /*0x01*/ u8 transitionFlags; // used to be named bike, but its definitely not that. seems to be some transition flags
     /*0x02*/ u8 runningState; // this is a static running state. 00 is not moving, 01 is turn direction, 02 is moving.
     /*0x03*/ u8 tileTransitionState; // this is a transition running state: 00 is not moving, 01 is transition between tiles, 02 means you are on the frame in which you have centered on a tile but are about to keep moving, even if changing directions. 2 is also used for a ledge hop, since you are transitioning.
-    /*0x04*/ u8 spriteId;
-    /*0x05*/ u8 objectEventId;
+    /*0x04*/ u32 spriteId;
+    /*0x05*/ u32 objectEventId;
     /*0x06*/ bool8 preventStep;
     /*0x07*/ u8 gender;
     /*0x08*/ u8 acroBikeState; // 00 is normal, 01 is turning, 02 is standing wheelie, 03 is hopping wheelie
